generic transfer element size
