+++ src/prelude.rs
@@ -1,6 +1,9 @@
 pub use afio::AfioExt as _stm32f103xx_hal_afio_AfioExt;
 pub use dma::DmaExt as _stm32f103xx_hal_dma_DmaExt;
 pub use dma::DmaChannel as _stm32f103xx_hal_dma_DmaChannel;
+pub use dma::CircReadDma as _stm32f103xx_hal_dma_CircReadDma;
+pub use dma::ReadDma as _stm32f103xx_hal_dma_ReadDma;
+pub use dma::WriteDma as _stm32f103xx_hal_dma_WriteDma;
 pub use flash::FlashExt as _stm32f103xx_hal_flash_FlashExt;
 pub use gpio::GpioExt as _stm32f103xx_hal_gpio_GpioExt;
 pub use hal::digital::StatefulOutputPin as _embedded_hal_digital_StatefulOutputPin;
@@ -8,7 +11,4 @@ pub use hal::digital::ToggleableOutputPin as _embedded_hal_digital_ToggleableOut
 pub use hal::prelude::*;
 pub use pwm::PwmExt as _stm32f103xx_hal_pwm_PwmExt;
 pub use rcc::RccExt as _stm32f103xx_hal_rcc_RccExt;
-pub use serial::CircReadDma as _stm32f103xx_hal_serial_CircReadDma;
-pub use serial::ReadDma as _stm32f103xx_hal_serial_ReadDma;
-pub use serial::WriteDma as _stm32f103xx_hal_serial_WriteDma;
 pub use time::U32Ext as _stm32f103xx_hal_time_U32Ext;
